# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_infer_api_v1_inference_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "language": {
            "default": "hi",
            "title": "Language",
            "type": "string"
          },
          "model_id": {
            "default": "aasist-v1",
            "title": "Model Id",
            "type": "string"
          },
          "upload_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Id"
          }
        },
        "title": "Body_infer_api_v1_inference_post",
        "type": "object"
      },
...
      "Body_upload_audio_api_v1_uploads_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_audio_api_v1_uploads_post",
        "type": "object"
      },
...
      "ExplainResponse": {
        "additionalProperties": false,
        "description": "Explainability artefacts.",
        "properties": {
          "artefacts": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Artefacts",
            "type": "array"
          }
        },
        "required": [
          "artefacts"
        ],
        "title": "ExplainResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PredictionResponse": {
        "additionalProperties": false,
        "description": "Inference verdict payload (REQ-087-091).",
        "properties": {
          "accent_notes": {
            "default": "",
            "title": "Accent Notes",
            "type": "string"
          },
          "analysis_summary": {
            "default": "",
            "title": "Analysis Summary",
            "type": "string"
          },
          "compression_status": {
            "title": "Compression Status",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language"
          },
          "enrichment_backend": {
            "default": "",
            "title": "Enrichment Backend",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "language_notes": {
            "default": "",
            "title": "Language Notes",
            "type": "string"
          },
          "prediction_id": {
            "title": "Prediction Id",
            "type": "string"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Probabilities",
            "type": "object"
          },
          "reliability": {
            "title": "Reliability",
            "type": "string"
          },
          "risk_notes": {
            "default": "",
            "title": "Risk Notes",
            "type": "string"
          },
          "spectrogram": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Spectrogram",
            "type": "array"
          },
          "transcript": {
            "default": "",
            "title": "Transcript",
            "type": "string"
          },
          "waveform": {
            "items": {
              "type": "number"
            },
            "title": "Waveform",
            "type": "array"
          },
          "whisper_backend": {
            "default": "",
            "title": "Whisper Backend",
            "type": "string"
          }
        },
        "required": [
          "prediction_id",
          "label",
          "confidence",
          "reliability",
          "language",
          "compression_status",
          "probabilities"
        ],
        "title": "PredictionResponse",
        "type": "object"
      },
...
      "ReportResponse": {
        "additionalProperties": false,
        "description": "Downloadable report pointer.",
        "properties": {
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "report_markdown": {
            "title": "Report Markdown",
            "type": "string"
          }
        },
        "required": [
          "report_markdown",
          "experiment_id"
        ],
        "title": "ReportResponse",
        "type": "object"
      },
      "UploadResponse": {
        "additionalProperties": false,
        "description": "Upload acknowledgement.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "upload_id": {
            "title": "Upload Id",
            "type": "string"
          }
        },
        "required": [
          "upload_id",
          "filename",
          "size_bytes",
          "content_type"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Cross-lingual, compression-robust AI-voice detection for Hindi, Marathi, and Tamil with calibrated reliability.",
    "title": "VaaniQ",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/experiments/report": {
      "get": {
        "description": "Downloadable evaluation report markdown.",
        "operationId": "experiment_report_api_v1_experiments_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "experiment_id",
            "required": false,
            "schema": {
              "default": "demo",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiment Report",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/report.md": {
      "get": {
        "description": "Raw markdown report download.",
        "operationId": "experiment_report_md_api_v1_experiments_report_md_get",
        "parameters": [
          {
            "in": "query",
            "name": "experiment_id",
            "required": false,
            "schema": {
              "default": "demo",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiment Report Md",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/explain": {
      "get": {
        "description": "Explainability artefacts for a prediction.",
        "operationId": "explain_api_v1_explain_get",
        "parameters": [
          {
            "in": "query",
            "name": "prediction_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain",
        "tags": [
          "explain"
        ]
      }
    },
...
    "/api/v1/inference": {
      "post": {
        "description": "Run inference from upload_id or multipart file (REQ-085-091).",
        "operationId": "infer_api_v1_inference_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_infer_api_v1_inference_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Infer",
        "tags": [
          "inference"
        ]
      }
    },
...
    "/api/v1/uploads": {
      "post": {
        "description": "Upload and validate an audio file (REQ-084, REQ-135).",
        "operationId": "upload_audio_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_api_v1_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Audio",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}